+++ R/utils.R
@@ -494,3 +494,19 @@ buildTabItem <- function (index, tabsetId, foundSelected, tabs = NULL, divTag = 
   }
   return(list(liTag = liTag, divTag = divTag))
 }
+
+
+
+
+validateIcon <- function (icon) 
+{
+  if (is.null(icon) || identical(icon, character(0))) {
+    return(icon)
+  }
+  else if (inherits(icon, "shiny.tag") && icon$name == "i") {
+    return(icon)
+  }
+  else {
+    stop("Invalid icon. Use Shiny's 'icon()' function to generate a valid icon")
+  }
+}